Use namedtuple in pre-defined error messages
In terms of readability, it might be confusing to see something like
PACKAGE_NOT_FOUND[0], because at first sight is not clear what it really
means.

To solve that, named tuples are used, thus making it possible to
access to a specific element using dot notation (for example:
PACKAGE_NOT_FOUND.error_type).